requirements.txt
asgiref==3.7.2
Django==4.2.6
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
phonenumbers==8.13.22
Pillow==10.0.1
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
